Fixing size overflow.
git-svn-id: https://svn.code.sf.net/p/gxemul/code2/gxemul/trunk@6271 5063f80b-35bb-49d0-95f4-ce7ec8300d17

+++ src/include/diskimage.h
@@ -68,9 +68,9 @@ struct diskimage {
 	struct diskimage_overlay *overlays;
 
 	int		chs_override;
-	int		cylinders;
+	int64_t		cylinders;
 	int		heads;
-	int		sectors_per_track;
+	int64_t		sectors_per_track;
 
 	int		rpms;
